Почти рабочее приложение без интерфейса
