+++ Def/Translator_Method.py
@@ -1,0 +1,7 @@
+from googletrans import Translator
+
+#Перевод фразы
+def translate_to_english(text):
+    translator = Translator()
+    translated_text = translator.translate(text, dest='en')
+    return translated_text.text
